requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.8.2
streamlit==1.37.1
requests==2.32.3
# advanced (install manually):
sentence-transformers
faiss-cpu